refactor(canvas): extract mcp app host
