+++ apps/canvas-workspace/src/renderer/src/components/mcp-apps/McpAppFrame/__tests__/McpAppFrameStyles.test.ts
@@ -1,7 +1,7 @@
 import { readFileSync } from 'fs';
 import { describe, expect, it } from 'vitest';
 
-const mcpAppCss = readFileSync(new URL('../McpAppFrame.css', import.meta.url), 'utf8');
+const mcpAppCss = readFileSync(new URL('../index.css', import.meta.url), 'utf8');
 
 describe('MCP App frame styles', () => {
   it('keeps the portaled inline surface above the Right Dock that contains chat', () => {
