+++ apps/canvas-workspace/src/renderer/src/components/mcp-apps/McpAppFrame/index.css
